Attempt precision change

+++ cirq-google/cirq_google/transformers/analytical_decompositions/two_qubit_to_sycamore.py
@@ -386,7 +386,7 @@ def _create_corrected_circuit(
     # Get the local equivalents
     b_0, b_1, a_0, a_1 = _find_local_equivalents(
         target_unitary,
-        program.unitary(qubit_order=cirq.QubitOrder.explicit([q0, q1]), dtype=target_unitary.dtype),
+        program.unitary(qubit_order=cirq.QubitOrder.explicit([q0, q1]), dtype=np.complex128),
     )
 
     # Apply initial corrections
